requirements.txt
# ─────────────────────────────────────────────────────────────
# FFIEC Call Report App — Single Requirements File
# Covers: backend + ai_agent + bigdata
# Install: pip install -r requirements.txt
# ─────────────────────────────────────────────────────────────

# ── Web framework ─────────────────────────────────────────────
fastapi
uvicorn
httpx
pydantic

# ── AI Agent ──────────────────────────────────────────────────
langchain
langchain-core
langchain-google-genai
langgraph
google-generativeai
sseclient-py

# ── Data / BigData ────────────────────────────────────────────
duckdb
pandas
pyarrow
boto3
pyspark==3.5.1
ffiec-data-collector
setuptools

# ── Utilities ─────────────────────────────────────────────────
requests
python-dotenv
